Parse the PNC query out of the AHO Xml
