+++ tests/helpers/generateMockPncQueryResultFromAho.ts
@@ -1,96 +1,7 @@
 import { XMLParser } from "fast-xml-parser"
-import parsePncDate from "src/lib/parsePncDate"
-import type { PncOffence, PncQueryResult } from "src/types/PncQueryResult"
-
-type OffenceDates = {
-  startDate: Date
-  endDate?: Date
-}
-type OffenceTimes = {
-  startTime?: string
-  endTime?: string
-}
-
-type AhoPncOffence = {
-  COF: {
-    "@_ACPOOffenceCode": string
-    "@_CJSOffenceCode": string
-    "@_IntfcUpdateType": string
-    "@_OffEndDate": string
-    "@_OffEndTime": string
-    "@_OffStartDate": string
-    "@_OffStartTime": string
-    "@_OffenceQualifier1": string
-    "@_OffenceQualifier2": string
-    "@_OffenceTitle": string
-    "@_ReferenceNumber": string
-  }
-}
-
-type AhoPncCourtCase = {
-  CCR: {
-    "@_CourtCaseRefNo": string
-    "@_CrimeOffenceRefNo": string
-    "@_IntfcUpdateType": string
-  }
-  Offences: {
-    Offence: AhoPncOffence | AhoPncOffence[]
-  }
-}
-
-export type ParsedAHO = {
-  AnnotatedHearingOutcome: {
-    CXE01: {
-      FSC: {
-        "@_FSCode": string
-        "@_IntfcUpdateType": string
-      }
-      IDS: {
-        "@_CRONumber": string
-        "@_Checkname": string
-        "@_IntfcUpdateType": string
-        "@_PNCID": string
-      }
-      CourtCases: {
-        CourtCase: AhoPncCourtCase | AhoPncCourtCase[]
-      }
-    }
-    PNCQueryDate: string
-  }
-}
-
-const extractDates = (offence: AhoPncOffence): OffenceDates => {
-  const startDate = parsePncDate(offence.COF["@_OffStartDate"])
-  if (!startDate) {
-    throw new Error(`Start date could not be processed: ${offence.COF["@_OffStartDate"]}`)
-  }
-  const dates: OffenceDates = {
-    startDate
-  }
-  if (offence.COF["@_OffEndDate"] && offence.COF["@_OffEndDate"] !== "") {
-    const endDate = parsePncDate(offence.COF["@_OffEndDate"])
-    if (endDate) {
-      dates.endDate = endDate
-    }
-  }
-
-  return dates
-}
-
-const extractTimes = (offence: AhoPncOffence): OffenceTimes => {
-  const times: OffenceTimes = { startTime: undefined, endTime: undefined }
-  const cofStartTime = offence.COF["@_OffStartTime"]
-  if (cofStartTime && cofStartTime !== "" && cofStartTime.length === 4) {
-    times.startTime = `${cofStartTime.substring(0, 2)}:${cofStartTime.substring(2)}`
-  }
-
-  const cofEndTime = offence.COF["@_OffEndTime"]
-  if (cofEndTime && cofEndTime !== "" && cofEndTime.length === 4) {
-    times.endTime = `${cofEndTime.substring(0, 2)}:${cofEndTime.substring(2)}`
-  }
-
-  return times
-}
+import mapXmlCxe01ToAho from "src/lib/mapXmlCxe01ToAho"
+import type { PncQueryResult } from "src/types/PncQueryResult"
+import type { Cxe01, RawAho } from "src/types/RawAho"
 
 /*
 Sample CXE element
@@ -114,51 +25,16 @@ Sample CXE element
 */
 
 export default (ahoXml: string): PncQueryResult | undefined => {
-  let cases
   const options = {
-    ignoreAttributes: false,
-    removeNSPrefix: true
+    ignoreAttributes: false
   }
   const parser = new XMLParser(options)
-  const rawParsedObj = parser.parse(ahoXml) as ParsedAHO
+  const rawParsedObj = parser.parse(ahoXml) as RawAho
 
-  const cxe = rawParsedObj.AnnotatedHearingOutcome.CXE01
-
+  const cxe = rawParsedObj["br7:AnnotatedHearingOutcome"]?.CXE01 as Cxe01
   if (!cxe) {
     return undefined
   }
 
-  if (cxe.CourtCases) {
-    if (!Array.isArray(cxe.CourtCases.CourtCase)) {
-      cxe.CourtCases.CourtCase = [cxe.CourtCases.CourtCase]
-    }
-    for (const courtCase of cxe.CourtCases.CourtCase) {
-      if (!Array.isArray(courtCase.Offences.Offence)) {
-        courtCase.Offences.Offence = [courtCase.Offences.Offence]
-      }
-    }
-    cases = cxe.CourtCases?.CourtCase.map((courtCase) => ({
-      courtCaseReference: courtCase.CCR["@_CourtCaseRefNo"],
-      offences: (courtCase.Offences.Offence as AhoPncOffence[]).map(
-        (offence): PncOffence => ({
-          offence: {
-            acpoOffenceCode: offence.COF["@_ACPOOffenceCode"],
-            cjsOffenceCode: offence.COF["@_CJSOffenceCode"],
-            sequenceNumber: parseInt(offence.COF["@_ReferenceNumber"], 10),
-            ...extractDates(offence),
-            ...extractTimes(offence)
-          }
-        })
-      )
-    }))
-  }
-
-  const checkName = cxe.IDS["@_Checkname"]
-  const result: PncQueryResult = {
-    forceStationCode: cxe.FSC["@_FSCode"],
-    checkName,
-    pncId: cxe.IDS["@_PNCID"],
-    cases
-  }
-  return result
+  return mapXmlCxe01ToAho(cxe)
 }
